+++ dflib/src/main/java/com/nhl/dflib/DataFrame.java
@@ -745,6 +745,24 @@ public interface DataFrame extends Iterable<RowProxy> {
      */
     DataFrame ne(DataFrame another);
 
+    /**
+     * Returns a new DataFrame with 3 columns "row", "column", "value", that contains values from all columns of
+     * this DataFrame. Null values are not included.
+     *
+     * @return a new DataFrame with columns called "row", "column", "value".
+     * @since 0.6
+     */
+    DataFrame stack();
+
+    /**
+     * Returns a new DataFrame with 3 columns "row", "column", "value", that contains values from all columns of
+     * this DataFrame. Null values are included.
+     *
+     * @return a new DataFrame with columns called "row", "column", "value".
+     * @since 0.6
+     */
+    DataFrame stackIncludeNulls();
+
     @Override
     Iterator<RowProxy> iterator();
 }