+++ dflib/src/main/java/com/nhl/dflib/concat/SeriesConcat.java
@@ -5,9 +5,11 @@ import com.nhl.dflib.Series;
 import com.nhl.dflib.series.ArraySeries;
 import com.nhl.dflib.series.IntArraySeries;
 
+import static java.util.Arrays.asList;
+
 public class SeriesConcat {
 
-    public static <T> Series<T> concat(Iterable<Series<T>> concat) {
+    public static <T> Series<T> concat(Series[] concat) {
         int h = 0;
         for (Series<? extends T> s : concat) {
             h += s.size();
@@ -22,6 +24,10 @@ public class SeriesConcat {
         }
 
         return new ArraySeries<>(data);
+    }
+
+    public static IntSeries intConcat(IntSeries[] concat) {
+        return intConcat(asList(concat));
     }
 
     public static IntSeries intConcat(Iterable<IntSeries> concat) {